# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.10.0
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    tr-engine authenticates **client software, not people**. Every client
    (a dashboard deployment, a script, an upload plugin, a website's
    backend, a phone app) holds its own **API key** and sends it as:

    ```
    Authorization: Bearer tre_<64 hex characters>
    ```

    There are no user accounts, logins, sessions or cookies. How a client
    authenticates its own users, if it has any, is up to the client. The
    operator guide and the client-developer guide are in `docs/auth.md`.

    **Scopes.** A key has one of `listen`, `edit` or `admin`, optionally
    plus `upload`, or only `upload`:

    | Scope | Grants | Implies |
    |---|---|---|
    | `listen` | Read radio data, stream events and audio, mint tickets | |
    | `edit` | Change radio metadata (tags, transcription corrections and reviews, unit-tag suggestions) | `listen` |
    | `admin` | Everything else: keys, anonymous access, identity edits, merges, maintenance, storage, SQL query, pages, imports, console logs, audit log | `edit`, `listen` |
...

...
    what applies to it.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /calls:
    get:
      operationId: listCalls
      x-scope: listen
      x-restricted: enforced
      summary: List calls
      description: |
        Returns call recordings with comprehensive filters. This is the
        primary endpoint for querying call history.

        **Common patterns:**
        - Live feed: `?sort=-stop_time&deduplicate=true&limit=50`
        - Talkgroup history: `?tgid=9178&sysid=348&start_time=...&end_time=...`
        - Unit activity: `?unit_id=924003`
        - Emergency calls: `?emergency=true&sort=-start_time`
      tags: [calls]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID (comma-separated for multiple, e.g., "348,34D"). Alias "sysids" also accepted.
          schema:
            type: string
        - name: system_id
          in: query
          description: Filter by system database ID (comma-separated for multiple). Alias "systems" also accepted.
          schema:
            type: string
            example: "1,2"
        - name: site_id
          in: query
          description: Filter by recording site ID (comma-separated for multiple). Alias "sites" also accepted.
          schema:
            type: string
            example: "1,3"
        - name: tgid
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple). Alias "tgids" also accepted.
          schema:
            type: string
            example: "9178,5344"
        - name: unit_id
          in: query
          description: Filter by unit radio ID (comma-separated for multiple). Aliases "units", "unit_ids" also accepted.
          schema:
            type: string
            example: "924003,104"
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - name: encrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
        - name: deduplicate
          in: query
          description: |
            When true, returns one call per call_group (best recording).
            Useful for feed views where simulcast duplicates are unwanted.
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - name: sort
          in: query
          description: |
            Sort field. Prefix with `-` for descending.
            Allowed: start_time, stop_time, duration, tgid, freq
          schema:
            type: string
            default: "-start_time"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/active:
    get:
      operationId: listActiveCalls
      x-scope: listen
      x-restricted: enforced
      summary: List active calls (real-time)
      description: |
        Returns currently active calls from the in-memory MQTT tracker.
        This is a live endpoint — it returns a point-in-time snapshot of
        all active calls from memory. There is no pagination; the full
        list is returned in a single response.

        Active calls may have incomplete fields (no stop_time, no audio_url,
        duration is elapsed-so-far). The response uses the same Call schema
        with nullable fields where data is not yet available.
      tags: [calls]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID
          schema:
            type: string
        - name: tgid
          in: query
          description: Filter by talkgroup ID
          schema:
            type: integer
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - name: encrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveCallListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{id}:
    get:
      operationId: getCall
      x-scope: listen
      x-restricted: enforced
      summary: Get a call
      description: |
        Returns a single call recording by its database ID.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{id}/audio:
    get:
      operationId: getCallAudio
      x-scope: listen
      x-restricted: enforced
      summary: Stream call audio
      description: |
        Streams the audio file for a call. Content-Type varies by source
        format. Responses carry `Cache-Control: private`.

        **Authentication:** `Authorization: Bearer <key>` when the client
        can send headers, `?ticket=<ticket>` from `POST /tickets` for an
        `<audio src>` element (which cannot), or no credential when the
        anonymous access policy is `listen`. An audio element makes
        repeated (Range) requests with the same URL; a ticket stays valid
        for all of them until it expires, so mint one with some lifetime
        left (for example at least 5 minutes) and mint a new one if
        playback fails with an error.

        A call outside a restricted credential's restriction is 404.

        `audio_url` fields in API responses hold this path and never a
        credential; see `Call.audio_url`.
      tags: [calls]
      security:
        - bearerAuth: []
        - ticketAuth: []
      parameters:
        - $ref: "#/components/parameters/callId"
        - $ref: "#/components/parameters/ticket"
      responses:
        "200":
          description: Audio stream
          headers:
            Cache-Control:
              schema:
                type: string
                example: private
          content:
            audio/mp4:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            audio/ogg:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calls/{id}/frequencies:
    get:
      operationId: getCallFrequencies
      x-scope: listen
      x-restricted: enforced
      summary: Get call frequency list
      description: |
        Returns the frequency entries (freqList) within a call, ordered by
        position in the audio.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallFrequencyListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{id}/transmissions:
    get:
      operationId: getCallTransmissions
      x-scope: listen
      x-restricted: enforced
      summary: Get call transmission list
      description: |
        Returns the unit transmissions (srcList) within a call, ordered by
        position in the audio. Each entry represents a unit keying up.
        Includes unit alpha_tags.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallTransmissionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Transcriptions
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: tre_<64 hex characters>
      description: |
        An API key in the `Authorization` header:
        `Authorization: Bearer tre_...`. Keys are created by an admin
        (`POST /keys` or `tr-engine keys create`). Keys are never accepted
        from the URL.
    ticketAuth:
      type: apiKey
      in: query
      name: ticket
      description: |
        A short-lived, listen-only ticket from `POST /tickets`, passed as
        `?ticket=trt_...`. Accepted only by `GET`/`HEAD` on
        `/events/stream`, `/audio/live` and `/calls/{id}/audio`, for
        browser APIs that cannot send headers (`EventSource`,
        `<audio src>`, `WebSocket`). Takes precedence over an
        `Authorization` header on those operations.

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
...
    ticket:
      name: ticket
      in: query
      required: false
      description: |
        A ticket from `POST /tickets`, for clients that cannot send an
        `Authorization` header (browser `EventSource`, `<audio>`,
        `WebSocket`). Takes precedence over the header. A key is never
        accepted here.
      schema:
        type: string
        example: trt_eyJrIjo3LCJlIjoxNzkwMDAwMDAwfQ.bWFjLWJ5dGVzLWhlcmU
    callId:
      name: id
      in: path
      required: true
      description: Call record database ID
      schema:
        type: integer
        example: 48531
...
    startTime:
      name: start_time
      in: query
      description: Start of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    endTime:
      name: end_time
      in: query
      description: End of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: |
        No usable credential. `code` is `key_required` (no credential, and
        the anonymous access policy doesn't allow this operation),
        `invalid_key` (unknown, revoked or expired key) or `invalid_ticket`
        (malformed, expired or orphaned ticket).
      headers:
        WWW-Authenticate:
          description: Always `Bearer realm="tr-engine"`
          schema:
            type: string
            example: Bearer realm="tr-engine"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            key_required:
              value:
                code: key_required
                error: an API key is required
            invalid_key:
              value:
                code: invalid_key
                error: API key has been revoked
    Forbidden:
      description: |
        The credential is valid but not allowed here. `code` is
        `insufficient_scope` (the message names the scope the operation
        needs) or `restricted_credential` (the credential is restricted to
        some systems or talkgroups, and this operation can't enforce that).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficient_scope:
              value:
                code: insufficient_scope
                error: this operation needs the edit scope
            restricted_credential:
              value:
                code: restricted_credential
                error: this operation is not available to restricted credentials
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
            - key_required
            - invalid_key
            - invalid_ticket
            - insufficient_scope
            - restricted_credential
          description: |
            Machine-readable error code for programmatic handling. Auth codes:
            - `key_required` (401): no credential, and the anonymous access
              policy doesn't allow this operation
            - `invalid_key` (401): unknown, revoked or expired key
            - `invalid_ticket` (401): malformed, expired or orphaned ticket,
              or one that references a merged-away system
            - `insufficient_scope` (403): the key lacks the needed scope
            - `restricted_credential` (403): the credential is restricted and
              this operation can't enforce that
            - `service_unavailable` (503): also when the credential lookup
              failed
            - `forbidden` (403): from the auth layer, only for a route with no
              auth policy (a server bug)
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    Call:
      type: object
      description: A recorded radio call/transmission
      required: [call_id, system_id, tgid, start_time]
      properties:
        # Identifiers
        call_id:
          type: integer
          description: Unique database ID for this recording
          example: 48531
        call_group_id:
          type: integer
          description: |
            Call group this recording belongs to. Multiple recordings of the
            same transmission (from different sites) share a call_group_id.
            Derived from (system_id, tgid, start_unix).
          example: 1

        # System context (embedded to avoid lookup)
        system_id:
          type: integer
          description: Logical radio system this call belongs to
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"

        # Site context (which recorder captured this)
        site_id:
          type: integer
          description: Recording site that captured this call
          example: 1
        site_short_name:
          type: string
          description: TR sys_name of the capturing site (e.g., "butco", "warco")
          example: butco

        # Talkgroup context (embedded to avoid lookup)
        tgid:
          type: integer
          example: 9178
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch
        tg_tag:
          type: string
          description: Functional category tag (e.g., Law Dispatch, Fire Tac)
          example: Law Dispatch
        tg_group:
          type: string
          description: Broad organizational grouping
          example: "Butler County (09) Law"

        # Timing
        start_time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        stop_time:
          type: string
          format: date-time
          nullable: true
          description: "Null for active (in-progress) calls"
          example: "2024-01-15T10:30:45Z"
        duration:
          type: number
          description: Duration in seconds (elapsed so far for active calls)
          example: 45.5

        # Audio
        audio_url:
          type: string
          nullable: true
          description: |
            Path of this call's audio on the engine origin
            (`/api/v1/calls/{id}/audio`), or null when there is no audio yet.
            It never contains a credential. Resolve it against the engine's
            base URL (not the page's origin, if the client is served from
            elsewhere), and, when the client authenticates with a key and
            cannot send headers (an `<audio>` element), append
            `?ticket=<ticket>` from `POST /tickets`.
          example: /api/v1/calls/48531/audio
        audio_type:
          type: string
          description: "Audio encoding type from trunk-recorder (e.g., `digital`, `digital tdma`, `analog`)"
          example: digital
        audio_size:
          type: integer
          description: Audio file size in bytes
          example: 45000

        # Signal quality
        freq:
          type: integer
          description: Frequency in Hz
          example: 851012500
        freq_error:
          type: integer
          description: Frequency error in Hz
          example: 0
        signal_db:
          type: number
          nullable: true
          description: Signal strength in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -23
        noise_db:
          type: number
          nullable: true
          description: Noise floor in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -50
        error_count:
          type: integer
          description: Decode error count
          example: 2
        spike_count:
          type: integer
          description: Audio spike count
          example: 0

        # Status
        call_state:
          $ref: "#/components/schemas/CallState"
        mon_state:
          $ref: "#/components/schemas/MonState"
        emergency:
          type: boolean
          example: false
        encrypted:
          type: boolean
          example: false
        analog:
          type: boolean
          example: false
        conventional:
          type: boolean
          example: false
        phase2_tdma:
          type: boolean
          example: false
        tdma_slot:
          type: integer
          description: "TDMA slot (0=slot 1 or N/A, 1=slot 2)"
          example: 0

        # Patches
        patched_tgids:
          type: array
          description: Other talkgroup IDs patched into this call
          items:
            type: integer

        # Transmission and frequency data (JSONB columns)
        src_list:
          type: array
          description: |
            Source/transmission list from trunk-recorder's srcList.
            Each entry is a unit transmission segment within the call.
            Also available via GET /calls/{id}/transmissions.
          items:
            $ref: "#/components/schemas/CallTransmission"
        freq_list:
          type: array
          description: |
            Frequency list from trunk-recorder's freqList.
            Each entry is a frequency segment within the call.
            Also available via GET /calls/{id}/frequencies.
          items:
            $ref: "#/components/schemas/CallFrequency"
        unit_ids:
          type: array
          description: Denormalized array of distinct unit radio IDs (src values) from src_list. GIN-indexed for fast unit filtering.
          items:
            type: integer

        # Units on this call (embedded)
        units:
          type: array
          items:
            $ref: "#/components/schemas/CallUnit"

        # Transcription (denormalized from transcriptions table)
        has_transcription:
          type: boolean
          description: Whether this call has been transcribed
          example: true
        transcription_status:
          $ref: "#/components/schemas/TranscriptionStatus"
        transcription_text:
          type: string
          nullable: true
          description: Primary transcription text (null if no transcription)
          example: Engine 5 responding to the scene
        transcription_word_count:
          type: integer
          nullable: true
          example: 6

        # Extensible metadata
        metadata_json:
          type: object
          additionalProperties: true
          description: Arbitrary metadata from trunk-recorder
...
    CallListResponse:
      type: object
      required: [calls, total, limit, offset]
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/Call"
        total:
          type: integer
          description: Total matching results (across all pages)
          example: 2500
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
    ActiveCallListResponse:
      type: object
      required: [calls, total]
      properties:
        calls:
          type: array
          description: |
            Active calls from MQTT tracker. Fields like stop_time and
            audio_url will be null since the call is still in progress.
          items:
            $ref: "#/components/schemas/Call"
        total:
          type: integer
          description: Number of currently active calls
          example: 5
...
    CallFrequencyListResponse:
      type: object
      required: [frequencies, total]
      properties:
        frequencies:
          type: array
          items:
            $ref: "#/components/schemas/CallFrequency"
        total:
          type: integer
          example: 3
...